Extract SongView legacy logic into hooks
